Se agrega un test se agrego test de prueba de llamada de los datos a la BD.

+++ src/main/java/com/projectT/projectTingeso/controller/StudentController.java
@@ -6,5 +6,5 @@ import javax.transaction.Transactional;
 
 
 @Transactional
-public interface StudentController extends CrudRepository<Student, Long> {
+public interface StudentController extends CrudRepository<Student, Integer> {
 }
